feature: add support for custom gateway
