+++ lib/handler.js
@@ -2,7 +2,7 @@
 /* eslint no-console: 0, indent: ["error", 4, { "SwitchCase": 1 }], max-len: [2, 150], max-statements: [2, 30], complexity: [0, 10] */
 
 const util = require('./util/validHandlers');
-const middleware = require('./middleware');
+const handlers = require('./middleware');
 
 /**
  * @description configure handlers
@@ -10,7 +10,7 @@ const middleware = require('./middleware');
  * @param {object} options user defined options
  * @returns {array} middleware collection
  */
-const handler = (route, options) => {
+const handler = (route, options = {}) => {
     const {
         schema,
         gateway,
@@ -29,8 +29,6 @@ const handler = (route, options) => {
         console.error(`Route ${route.url} is misconfigured`);
         throw new Error('Request Schema contains custom or locals with no properties. Please check the route configuration.');
     }
-
-    const handlers = options ? { ...middleware, ...options } : middleware;
 
     // NOTE: this order is critical to run application as expected.
     // DO NOT CHANGE middleware order
@@ -74,7 +72,8 @@ const handler = (route, options) => {
             }
             case 'gateway': {
                 if (gateway && util.isValidGateway(gateway)) {
-                    acc.push(fn(gateway));
+                    const { gateway: customGateway } = options;
+                    acc.push(fn(gateway, customGateway));
                 }
                 break;
             }
