+++ lib/middleware/filter.js
@@ -1,3 +1,4 @@
+const _ = require('lodash');
 const responseSchema = require('../response/responseSchema');
 
 /**
@@ -8,10 +9,13 @@ const responseSchema = require('../response/responseSchema');
  * @returns {any} releases control to next middleware
  */
 const param = () => (req, res, next) => {
-    const { routerConfig: { schema: { response: { ok } } } } = req;
-    const data = res.props('data');
-    const result = responseSchema(ok, data, res);
-    res.setProps('data', result);
+    const ok = _.get(req, 'routerConfig.schema.response.ok');
+    if (ok) {
+        const data = res.props('data');
+        const result = responseSchema(ok, data, res);
+        res.setProps('data', result);
+    }
+
     return next();
 };
