+++ lib/middleware/method.js
@@ -10,7 +10,7 @@ const {
 } = require('./config');
 
 module.exports = route => function attachConfig(req, res, next) {
-    res.locals[RESPONSE_CUSTOM_PROP_KEY] = {
+    res[RESPONSE_CUSTOM_PROP_KEY] = {
         [RESPONSE_CUSTOM_PROP_DATA]: {},
         [RESPONSE_CUSTOM_PROP_REQ_PARAMS]: null,
         [RESPONSE_CUSTOM_PROP_REQ_CONFIG]: null,
@@ -18,9 +18,9 @@ module.exports = route => function attachConfig(req, res, next) {
         [RESPONSE_CUSTOM_PROP_REDIRECT]: null
     };
     // add prop getter and setter
-    res.props = key => _.get(res.locals[RESPONSE_CUSTOM_PROP_KEY], key);
+    res.props = key => _.get(res[RESPONSE_CUSTOM_PROP_KEY], key);
     res.setProps = (key, value) => {
-        res.locals[RESPONSE_CUSTOM_PROP_KEY][key] = value;
+        res[RESPONSE_CUSTOM_PROP_KEY][key] = value;
     };
 
     req.routerConfig = route;
